+++ src/CHANGELOG.md
@@ -10,7 +10,8 @@ Resulted image will respect filtered types and multi-z rendering if enabled.
 Tool has a quick shortcut to toggle: hold R key.
 
 ### Improvements and Fixes
- * Fixed self-update for Windows when StrongDMM executable was not on the C drive.
+ * Fixed self-update for Windows when StrongDMM executable was not on the C drive;
+ * Unnecessary updates can now be ignored.
 
 # v2.2.0.alpha